pyproject.toml
[tool.poetry]
name = "gcalls"
version = "0.1.0"
description = ""
authors = ["Clement Walter <clement0walter@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9, <3.10"
pandas = "^1.3.2"
inflection = "^0.5.1"
gspread = "^4.0.1"
oauth2client = "^4.1.3"
APScheduler = "^3.7.0"
numpy = "^1.21.2"
python-dotenv = "^0.20.0"
starknet-py = "^0.15.2"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
ipython = "^7.27.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"
starknet-devnet = "0.5.0a2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git
    | \.pytest_cache
    | \.venv
    | build
    | dist
    | libs
  )/
)
'''

[tool.isort]
profile = "black"
atomic = true
include_trailing_comma = true
use_parentheses = true
